requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
Faker==37.3.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
